requirements.txt
streamlit
Pillow
google-generativeai
supabase